pyproject.toml
[project]
name = "agentic-rag-app"
version = "0.1.0"
description = "An agentic RAG application using LangGraph, Chainlit, ChromaDB, and Gemini."
authors = [
    {name = "Praveen Kumar Verma", email = "praveen@example.com"},
]
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.6.9",
    "langchain>=0.3.27",
    "langgraph>=0.6.5",
    "langchain-google-genai>=2.1.9",
    "langchain-huggingface>=0.3.1",
    "sentence-transformers>=5.1.0",
    "torch>=2.0.0",
    "chromadb>=1.0.17",
    "pypdf>=6.0.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.32.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.1.1",
    "langchain-community>=0.3.27",
    "langchain-text-splitters>=0.3.9",
]

# For CPU-only PyTorch installation, use:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

[tool.pip]
# This will hint package managers to prefer CPU versions
extra-index-url = [
    "https://download.pytorch.org/whl/cpu"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
